Finished The Starting Hurdles
It now does the following.  On a new member arvial it asigns 'Initiate' role to the user and then asks them to provide their IGN.  After they type their IGN their role is changed from 'Initiate' to 'Shadow Soldier' and the bot then lays out our rules and the rest of the commands to use (which aren't in place yet) that access the database.