Added repeat offender function
Using Redis I have it count each time a particular offense occurs.  The key is created from the username so it is unique and easily retrievable.  At different amount of offenses I can have it post different messages or change the users role or even kick/ban if it gets bad enough.